% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_ml.R
\name{pbetamp}
\alias{pbetamp}
\title{Beta distribution function, mean-precision parametrization}
\usage{
pbetamp(y, mu, phi)
}
\arguments{
\item{y}{Numeric vector in \verb{(0, 1)}.}

\item{mu}{Numeric vector, the mean, in \verb{(0, 1)}.}

\item{phi}{Numeric vector, the precision parameter, \verb{> 0}.}
}
\value{
Numeric vector, \code{F(y)}.
}
\description{
Beta distribution function, mean-precision parametrization
}
